docs: Extended API
