+++ cognite/pygen/utils/text.py
@@ -7,12 +7,21 @@ import inflect
 
 def to_camel(string: str, pluralize: bool = False, singularize: bool = False) -> str:
     """Convert snake_case_name to camelCaseName.
-    >>> to_camel("a_b")
-    'aB'
-    >>> to_camel('camel_case', pluralize=True)
-    'camelCases'
-    >>> to_camel('best_directors', singularize=True)
-    'bestDirector'
+
+    Args:
+        string: The string to convert.
+        pluralize: Whether to pluralize the last word.
+        singularize: Whether to singularize the last word.
+    Returns:
+        camelCase of the input string.
+
+    Examples:
+        >>> to_camel("a_b")
+        'aB'
+        >>> to_camel('camel_case', pluralize=True)
+        'camelCases'
+        >>> to_camel('best_directors', singularize=True)
+        'bestDirector'
     """
     if "_" in string:
         # Is snake case
@@ -37,19 +46,28 @@ def to_camel(string: str, pluralize: bool = False, singularize: bool = False) ->
 
 
 def to_pascal(string: str, pluralize: bool = False, singularize: bool = False) -> str:
-    """Convert snake_case_name to PascalCaseName.
-    >>> to_pascal("a_b")
-    'AB'
-    >>> to_pascal('camel_case', pluralize=True)
-    'CamelCases'
-    >>> to_pascal('best_directors', singularize=True)
-    'BestDirector'
-    >>> to_pascal("BestLeadingActress", singularize=True)
-    'BestLeadingActress'
-    >>> to_pascal("priceScenarios", pluralize=True)
-    'PriceScenarios'
-    >>> to_pascal("reserveScenarios", pluralize=True)
-    'ReserveScenarios'
+    """Convert string to PascalCaseName.
+
+    Args:
+        string: The string to convert.
+        pluralize: Whether to pluralize the last word.
+        singularize: Whether to singularize the last word.
+    Returns:
+        PascalCase of the input string.
+
+    Examples:
+        >>> to_pascal("a_b")
+        'AB'
+        >>> to_pascal('camel_case', pluralize=True)
+        'CamelCases'
+        >>> to_pascal('best_directors', singularize=True)
+        'BestDirector'
+        >>> to_pascal("BestLeadingActress", singularize=True)
+        'BestLeadingActress'
+        >>> to_pascal("priceScenarios", pluralize=True)
+        'PriceScenarios'
+        >>> to_pascal("reserveScenarios", pluralize=True)
+        'ReserveScenarios'
     """
     camel = to_camel(string, pluralize, singularize)
     return f"{camel[0].upper()}{camel[1:]}" if camel else ""
@@ -57,51 +75,60 @@ def to_pascal(string: str, pluralize: bool = False, singularize: bool = False) -
 
 def to_snake(string: str, pluralize: bool = False, singularize: bool = False) -> str:
     """
-    Convert PascalCaseName to snake_case_name.
-    >>> to_snake("aB")
-    'a_b'
-    >>> to_snake('CamelCase')
-    'camel_case'
-    >>> to_snake('camelCamelCase')
-    'camel_camel_case'
-    >>> to_snake('Camel2Camel2Case')
-    'camel_2_camel_2_case'
-    >>> to_snake('getHTTPResponseCode')
-    'get_http_response_code'
-    >>> to_snake('get200HTTPResponseCode')
-    'get_200_http_response_code'
-    >>> to_snake('getHTTP200ResponseCode')
-    'get_http_200_response_code'
-    >>> to_snake('HTTPResponseCode')
-    'http_response_code'
-    >>> to_snake('ResponseHTTP')
-    'response_http'
-    >>> to_snake('ResponseHTTP2')
-    'response_http_2'
-    >>> to_snake('Fun?!awesome')
-    'fun_awesome'
-    >>> to_snake('Fun?!Awesome')
-    'fun_awesome'
-    >>> to_snake('10CoolDudes')
-    '10_cool_dudes'
-    >>> to_snake('20coolDudes')
-    '20_cool_dudes'
-    >>> to_snake('BestDirector', pluralize=True)
-    'best_directors'
-    >>> to_snake('BestDirectors', singularize=True)
-    'best_director'
-    >>> to_snake('BestLeadingActress', pluralize=True)
-    'best_leading_actresses'
-    >>> to_snake('APM_Activity', pluralize=True)
-    'apm_activities'
-    >>> to_snake('APM_Activities', singularize=True)
-    'apm_activity'
-    >>> to_snake('APM_Operation', pluralize=True)
-    'apm_operations'
-    >>> to_snake('APM_Asset', pluralize=True)
-    'apm_assets'
-    >>> to_snake('APM_Material', pluralize=True)
-    'apm_materials'
+    Convert input string to snake_case
+
+    Args:
+        string: The string to convert.
+        pluralize: Whether to pluralize the last word.
+        singularize: Whether to singularize the last word.
+    Returns:
+        snake_case of the input string.
+
+    Examples:
+        >>> to_snake("aB")
+        'a_b'
+        >>> to_snake('CamelCase')
+        'camel_case'
+        >>> to_snake('camelCamelCase')
+        'camel_camel_case'
+        >>> to_snake('Camel2Camel2Case')
+        'camel_2_camel_2_case'
+        >>> to_snake('getHTTPResponseCode')
+        'get_http_response_code'
+        >>> to_snake('get200HTTPResponseCode')
+        'get_200_http_response_code'
+        >>> to_snake('getHTTP200ResponseCode')
+        'get_http_200_response_code'
+        >>> to_snake('HTTPResponseCode')
+        'http_response_code'
+        >>> to_snake('ResponseHTTP')
+        'response_http'
+        >>> to_snake('ResponseHTTP2')
+        'response_http_2'
+        >>> to_snake('Fun?!awesome')
+        'fun_awesome'
+        >>> to_snake('Fun?!Awesome')
+        'fun_awesome'
+        >>> to_snake('10CoolDudes')
+        '10_cool_dudes'
+        >>> to_snake('20coolDudes')
+        '20_cool_dudes'
+        >>> to_snake('BestDirector', pluralize=True)
+        'best_directors'
+        >>> to_snake('BestDirectors', singularize=True)
+        'best_director'
+        >>> to_snake('BestLeadingActress', pluralize=True)
+        'best_leading_actresses'
+        >>> to_snake('APM_Activity', pluralize=True)
+        'apm_activities'
+        >>> to_snake('APM_Activities', singularize=True)
+        'apm_activity'
+        >>> to_snake('APM_Operation', pluralize=True)
+        'apm_operations'
+        >>> to_snake('APM_Asset', pluralize=True)
+        'apm_assets'
+        >>> to_snake('APM_Material', pluralize=True)
+        'apm_materials'
     """
     pattern = re.compile(r"[A-Z]?[a-z]+|[A-Z]+(?=[A-Z][a-z]|\d|\W|$)|\d+")
     if "_" in string:
@@ -139,12 +166,19 @@ class _Inflect:
 
 def as_plural(noun: str) -> str:
     """Pluralize a noun.
-    >>> as_plural('person')
-    'persons'
-    >>> as_plural('Roles')
-    'Roles'
-    >>> as_plural('activity')
-    'activities'
+
+    Args:
+        noun: The noun to pluralize.
+    Returns:
+        The pluralized noun.
+
+    Examples:
+        >>> as_plural('person')
+        'persons'
+        >>> as_plural('Roles')
+        'Roles'
+        >>> as_plural('activity')
+        'activities'
     """
     if noun.lower() in _S_EXCEPTIONS:
         return f"{noun}s" if noun[-1] != "s" else noun
@@ -159,12 +193,19 @@ def as_plural(noun: str) -> str:
 
 def as_singular(noun: str) -> str:
     """Singularize a noun.
-    >>> as_singular('persons')
-    'person'
-    >>> as_singular('Roles')
-    'Role'
-    >>> as_singular('role')
-    'role'
+
+    Args:
+        noun: The noun to singularize.
+    Returns:
+        The singularized noun.
+
+    Examples:
+        >>> as_singular('persons')
+        'person'
+        >>> as_singular('Roles')
+        'Role'
+        >>> as_singular('role')
+        'role'
     """
     if noun.lower() in _S_EXCEPTIONS:
         return noun[:-1] if noun[-1] == "s" else noun
