+++ docs/api/utils_cdf.md
@@ -1,0 +1,1 @@
+::: cognite.pygen.utils.cdf
