+++ docs/api/utils_text.md
@@ -1,0 +1,1 @@
+::: cognite.pygen.utils.text
